refactor: Editor 페이지의 로직을 하위 컴포넌트로 옮김 #363
수정, 삭제 시 요청을 하위 컴포넌트로 옮김
