+++ frontend/src/components/Edit/Header.tsx
@@ -1,11 +1,19 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
 import EmojiPicker from 'emoji-picker-react';
 
 import useEditStore from '@store/useEditStore';
 
 const Header = () => {
+  const { state } = useLocation();
   const [showEmoji, setShowEmoji] = useState(false);
   const { title, setTitle, emoji, setEmoji, status, setStatus } = useEditStore();
+
+  useEffect(() => {
+    setTitle(state?.title || '');
+    setEmoji(state?.emoji || '😁');
+    setStatus(state?.status === 'public' ? '공개 하기' : '나만 보기');
+  }, [state]);
 
   const toggleEmoji = () => setShowEmoji((prev) => !prev);
